+++ src/trtutils/impls/yolo/_process.py
@@ -9,12 +9,12 @@ import cv2
 import numpy as np
 from cv2ext.image import letterbox
 
-from trtutils.impls.common import decode_efficient_nms
+from trtutils.impls.common import decode_efficient_nms, postprocess_efficient_nms
 
-from ._version import VALID_VERSIONS
+# EfficientNMS as 4 outputs
+_EFF_NUM_OUTPUTS = 4
 
 _log = logging.getLogger(__name__)
-_VERSION_CUTOFF = 10
 
 
 def preprocess(
@@ -44,11 +44,9 @@ def preprocess(
 
     tensor, ratios, padding = letterbox(image, new_shape=input_shape)
     tensor = cv2.cvtColor(tensor, cv2.COLOR_BGR2RGB)
-    # tensor = tensor.transpose((2, 0, 1))
-    # tensor = np.expand_dims(tensor, 0)
-    # tensor = tensor / 255.0
 
     tensor = tensor / 255.0  # type: ignore[assignment]
+
     tensor = tensor[np.newaxis, :]
     tensor = np.transpose(tensor, (0, 3, 1, 2))
     # large performance hit to assemble contiguous array
@@ -57,29 +55,6 @@ def preprocess(
     _log.debug(f"Ratios: {ratios}")
     _log.debug(f"Padding: {padding}")
     return tensor, ratios, padding
-
-
-def _postprocess_v_7_8_9(
-    outputs: list[np.ndarray],
-    ratios: tuple[float, float],
-    padding: tuple[float, float],
-) -> list[np.ndarray]:
-    # efficient NMS postprocessor essentially
-    # inputs are list[num_dets, bboxes, scores, classes]
-    num_dets, bboxes, scores, class_ids = outputs
-    ratio_width, ratio_height = ratios
-    pad_x, pad_y = padding
-
-    n_boxes = len(bboxes) // 4
-    adjusted_bboxes = bboxes.copy().reshape(n_boxes, 4)
-    adjusted_bboxes[:, 0] = (adjusted_bboxes[:, 0] - pad_x) / ratio_width  # x1
-    adjusted_bboxes[:, 1] = (adjusted_bboxes[:, 1] - pad_y) / ratio_height  # y1
-    adjusted_bboxes[:, 2] = (adjusted_bboxes[:, 2] - pad_x) / ratio_width  # x2
-    adjusted_bboxes[:, 3] = (adjusted_bboxes[:, 3] - pad_y) / ratio_height  # y2
-
-    adjusted_bboxes = np.clip(adjusted_bboxes, 0, None)
-
-    return [num_dets, adjusted_bboxes, scores, class_ids]
 
 
 def _postprocess_v_10(
@@ -115,7 +90,6 @@ def _postprocess_v_10(
 
 def postprocess(
     outputs: list[np.ndarray],
-    version: int,
     ratios: tuple[float, float] = (1.0, 1.0),
     padding: tuple[float, float] = (0.0, 0.0),
 ) -> list[np.ndarray]:
@@ -126,8 +100,6 @@ def postprocess(
     ----------
     outputs : list[np.ndarray]
         The outputs from a YOLO network.
-    version : int
-        The version of the YOLO networks.
     ratios : tuple[float, float]
         The ratio of original image to preprocessed shape
     padding : tuple[float, float]
@@ -139,16 +111,9 @@ def postprocess(
         The postprocessed outputs.
 
     """
-    if version < _VERSION_CUTOFF:
-        return _postprocess_v_7_8_9(outputs, ratios, padding)
+    if len(outputs) == _EFF_NUM_OUTPUTS:
+        return postprocess_efficient_nms(outputs, ratios, padding)
     return _postprocess_v_10(outputs, ratios, padding)
-
-
-def _get_detections_v_7_8_9(
-    outputs: list[np.ndarray],
-    conf_thres: float | None = None,
-) -> list[tuple[tuple[int, int, int, int], float, int]]:
-    return decode_efficient_nms(outputs, conf_thres=conf_thres)
 
 
 def _get_detections_v_10(
@@ -180,7 +145,6 @@ def _get_detections_v_10(
 
 def get_detections(
     outputs: list[np.ndarray],
-    version: int,
     conf_thres: float | None = None,
 ) -> list[tuple[tuple[int, int, int, int], float, int]]:
     """
@@ -190,8 +154,6 @@ def get_detections(
     ----------
     outputs : list[np.ndarray]
         The outputs from a YOLO networks.
-    version : int
-        Which version of YOLO used to generate the outputs.
     conf_thres : float, optional
         The confidence threshold to use for getting detections.
 
@@ -201,22 +163,7 @@ def get_detections(
         The detections from the YOLO netowrk.
         Each detection is a bounding box in form x1, y1, x2, y2, a confidence score and a class id.
 
-    Raises
-    ------
-    ValueError
-        If the version provided is invalid
-        If version is V10 and image width/height are not provided
-
     """
-    if version not in VALID_VERSIONS:
-        err_msg = (
-            f"Invalid version provided. Found: {version}, not in: {VALID_VERSIONS}"
-        )
-        raise ValueError(err_msg)
-    # Handle YoloV 7/8/9
-    if version < _VERSION_CUTOFF:
-        return _get_detections_v_7_8_9(outputs, conf_thres=conf_thres)
-    # Handle YoloV10
-    if conf_thres is None:
-        return _get_detections_v_10(outputs, conf_thres=conf_thres)
+    if len(outputs) == _EFF_NUM_OUTPUTS:
+        return decode_efficient_nms(outputs, conf_thres=conf_thres)
     return _get_detections_v_10(outputs, conf_thres=conf_thres)