+++ src/trtutils/impls/yolo/_yolos.py
@@ -1,0 +1,98 @@
+# Copyright (c) 2024 Justin Davis (davisjustin302@gmail.com)
+#
+# MIT License
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
+from ._yolo import YOLO
+
+if TYPE_CHECKING:
+    from pathlib import Path
+
+    from typing_extensions import Self
+
+
+class YOLOX(YOLO):
+    """Implementation of YOLOX."""
+
+    def __init__(
+        self: Self,
+        engine_path: Path | str,
+        warmup_iterations: int = 10,
+        *,
+        warmup: bool | None = None,
+    ) -> None:
+        super().__init__(
+            engine_path=engine_path,
+            warmup_iterations=warmup_iterations,
+            warmup=warmup,
+        )
+
+
+class YOLO7(YOLO):
+    """Implementation of YOLO7."""
+
+    def __init__(
+        self: Self,
+        engine_path: Path | str,
+        warmup_iterations: int = 10,
+        *,
+        warmup: bool | None = None,
+    ) -> None:
+        super().__init__(
+            engine_path=engine_path,
+            warmup_iterations=warmup_iterations,
+            warmup=warmup,
+        )
+
+
+class YOLO8(YOLO):
+    """Implementation of YOLO8."""
+
+    def __init__(
+        self: Self,
+        engine_path: Path | str,
+        warmup_iterations: int = 10,
+        *,
+        warmup: bool | None = None,
+    ) -> None:
+        super().__init__(
+            engine_path=engine_path,
+            warmup_iterations=warmup_iterations,
+            warmup=warmup,
+        )
+
+
+class YOLO9(YOLO):
+    """Implementation of YOLO9."""
+
+    def __init__(
+        self: Self,
+        engine_path: Path | str,
+        warmup_iterations: int = 10,
+        *,
+        warmup: bool | None = None,
+    ) -> None:
+        super().__init__(
+            engine_path=engine_path,
+            warmup_iterations=warmup_iterations,
+            warmup=warmup,
+        )
+
+
+class YOLO10(YOLO):
+    """Implementation of YOLO10."""
+
+    def __init__(
+        self: Self,
+        engine_path: Path | str,
+        warmup_iterations: int = 10,
+        *,
+        warmup: bool | None = None,
+    ) -> None:
+        super().__init__(
+            engine_path=engine_path,
+            warmup_iterations=warmup_iterations,
+            warmup=warmup,
+        )