+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 0.3.2 (10-31-2024)
+
+### Added
+
+- impls.yolo.YOLO variations:
+    YOLO7, YOLO8, YOLO9, YOLO10, YOLOX
+    All aliases for YOLO
+
+### Changed
+
+- impls.yolo.YOLO no longer requires a version to be given.
+    Automatically infers postprocessing from outputs.
+
 ## 0.3.1 (10-29-2024)
 
 ### Improved